+++ pyproject.toml
@@ -26,6 +26,5 @@ dev = [
     "pytest>=8.3.4",
     "ruff>=0.9.3",
 ]
-
-[tool.uv.workspace]
-members = ["packages/*"]
+[tool.uv]
+package = true
