+++ src/dynamic_simulation/__init__.py
@@ -1,0 +1,1 @@
+from .kite_system import KiteSystem
