+++ src/dynamic_simulation/benchmark/cylinder.urdf
