+++ src/dynamic_simulation/benchmark/mujoco_cable_system.xml
