+++ src/dynamic_simulation/benchmark/sphere.urdf
