+++ src/dynamic_simulation/exudyn_tethered_foil.py
@@ -18,7 +18,7 @@ import exudyn as exu
 import numpy as np
 from exudyn.rigidBodyUtilities import RotXYZ2RotationMatrix
 
-from dynamic_simulation.kite_system import KiteSystem
+from src.dynamic_simulation.kite_system import KiteSystem
 
 # Example usage with visualization and rendering:
