+++ src/dynamic_simulation/kite_system.py
